Added code to customerApp and Customer DB to help move forward with the Database and test the DB
